🌱 Split GPU and cluster component modules
Splits large GPU and cluster components into focused hooks and
sub-components, following the pattern used by useGPUReservationForm.ts
(#21717).

- Clusters.tsx (599 -> 411 lines): extracted useClusterViewState
  (filter/sort/layout state + URL sync), useClusterMutations
  (rename/remove network calls), and clusterStatValues.ts (pure
  StatsOverview value mapping).
- NamespaceResources.tsx (660 -> 494 lines): extracted
  namespaceResourceUtils.ts (buildAllResources, getStatusBgColor),
  ResourceTypeAccordion and SimpleResourceRow components to remove
  eight near-identical accordion/row blocks.
- GPUDetailModal.tsx (506 -> 197 lines): extracted gpuDetailUtils.ts
  (pure calculations), GPUSpecsPanel, GPUUtilizationList, and
  GPUNodesTable components.
- ReservationFormModal.tsx (734 -> 237 lines): extracted
  useReservationFormState (all form state, validation, and the
  save/quota-provisioning workflow) plus ClusterPicker,
  ResourceRequestFields, and ScheduleSelector field components.

No behavior changes — all logic preserved verbatim in the extracted
modules.

Fixes #21505
Fixes #21613
Fixes #21617
