+++ web/src/components/clusters/components/GPUDetailModal.tsx
@@ -1,10 +1,19 @@
 import { useMemo } from 'react'
-import { Zap, Server, Layers, RefreshCw, Cpu, AlertCircle, HardDrive, CircuitBoard, Settings } from 'lucide-react'
+import { Zap, Server, Layers, RefreshCw, Cpu, AlertCircle, CircuitBoard, Settings } from 'lucide-react'
 import { GPUNode, NVIDIAOperatorStatus } from '../../../hooks/useMCP'
 import { BaseModal } from '../../../lib/modals'
 import { useTranslation } from 'react-i18next'
-import { StatusBadge } from '../../ui/StatusBadge'
-import { wrapAbbreviations } from '../../shared/TechnicalAcronym'
+import {
+  computeGpuTypeInfo,
+  computeClusterInfo,
+  computeGpuTotals,
+  computeManufacturerBreakdown,
+  computeGpuSpecs,
+  getUtilizationColor,
+} from './gpuDetailUtils'
+import { GPUSpecsPanel, NvidiaOperatorStatusPanel } from './GPUSpecsPanel'
+import { GPUTypeUtilizationList, GPUClusterUtilizationList } from './GPUUtilizationList'
+import { GPUNodesTable } from './GPUNodesTable'
 
 interface GPUDetailModalProps {
   gpuNodes: GPUNode[]
@@ -15,165 +24,21 @@ interface GPUDetailModalProps {
   operatorStatus?: NVIDIAOperatorStatus[]
 }
 
-interface GPUTypeInfo {
-  type: string
-  manufacturer: string
-  totalGPUs: number
-  allocatedGPUs: number
-  availableGPUs: number
-  nodeCount: number
-  clusters: string[]
-}
-
-interface ClusterGPUInfo {
-  cluster: string
-  totalGPUs: number
-  allocatedGPUs: number
-  availableGPUs: number
-  nodeCount: number
-  gpuTypes: string[]
-}
-
-function extractManufacturer(gpuType: string): string {
-  const lower = gpuType.toLowerCase()
-  if (lower.includes('nvidia')) return 'NVIDIA'
-  if (lower.includes('amd') || lower.includes('radeon')) return 'AMD'
-  if (lower.includes('intel')) return 'Intel'
-  return 'Unknown'
-}
-
-const GPU_UTIL_HIGH = 90 // Utilization % threshold for critical (red) status
-const GPU_UTIL_WARN = 70 // Utilization % threshold for warning (yellow) status
-
-function getUtilizationColor(percentage: number): string {
-  if (percentage >= GPU_UTIL_HIGH) return 'text-red-400'
-  if (percentage >= GPU_UTIL_WARN) return 'text-yellow-400'
-  return 'text-green-400'
-}
-
 interface GPUDetailModalInternalProps extends GPUDetailModalProps {
   isOpen?: boolean
 }
 
 export function GPUDetailModal({ isOpen = true, gpuNodes, isLoading, error, onRefresh, onClose, operatorStatus }: GPUDetailModalInternalProps) {
-
   const { t } = useTranslation()
-  // Calculate GPU type breakdown
-  const gpuTypeInfo = useMemo(() => {
-    const typeMap = new Map<string, GPUTypeInfo>()
 
-    gpuNodes.forEach(node => {
-      const existing = typeMap.get(node.gpuType)
-      if (existing) {
-        existing.totalGPUs += node.gpuCount
-        existing.allocatedGPUs += node.gpuAllocated
-        existing.availableGPUs += (node.gpuCount - node.gpuAllocated)
-        existing.nodeCount += 1
-        if (!existing.clusters.includes(node.cluster)) {
-          existing.clusters.push(node.cluster)
-        }
-      } else {
-        typeMap.set(node.gpuType, {
-          type: node.gpuType,
-          manufacturer: extractManufacturer(node.gpuType),
-          totalGPUs: node.gpuCount,
-          allocatedGPUs: node.gpuAllocated,
-          availableGPUs: node.gpuCount - node.gpuAllocated,
-          nodeCount: 1,
-          clusters: [node.cluster] })
-      }
-    })
+  const gpuTypeInfo = useMemo(() => computeGpuTypeInfo(gpuNodes), [gpuNodes])
+  const clusterInfo = useMemo(() => computeClusterInfo(gpuNodes), [gpuNodes])
+  const gpuSpecs = useMemo(() => computeGpuSpecs(gpuNodes), [gpuNodes])
+  const totals = computeGpuTotals(gpuNodes)
+  const manufacturerBreakdown = computeManufacturerBreakdown(gpuTypeInfo)
 
-    return Array.from(typeMap.values()).sort((a, b) => b.totalGPUs - a.totalGPUs)
-  }, [gpuNodes])
-
-  // Calculate cluster breakdown
-  const clusterInfo = useMemo(() => {
-    const clusterMap = new Map<string, ClusterGPUInfo>()
-
-    gpuNodes.forEach(node => {
-      const existing = clusterMap.get(node.cluster)
-      if (existing) {
-        existing.totalGPUs += node.gpuCount
-        existing.allocatedGPUs += node.gpuAllocated
-        existing.availableGPUs += (node.gpuCount - node.gpuAllocated)
-        existing.nodeCount += 1
-        if (!existing.gpuTypes.includes(node.gpuType)) {
-          existing.gpuTypes.push(node.gpuType)
-        }
-      } else {
-        clusterMap.set(node.cluster, {
-          cluster: node.cluster,
-          totalGPUs: node.gpuCount,
-          allocatedGPUs: node.gpuAllocated,
-          availableGPUs: node.gpuCount - node.gpuAllocated,
-          nodeCount: 1,
-          gpuTypes: [node.gpuType] })
-      }
-    })
-
-    return Array.from(clusterMap.values()).sort((a, b) => b.totalGPUs - a.totalGPUs)
-  }, [gpuNodes])
-
-  // Calculate totals
-  const totals = (() => {
-    let total = 0
-    let allocated = 0
-    gpuNodes.forEach(node => {
-      total += node.gpuCount
-      allocated += node.gpuAllocated
-    })
-    return {
-      total,
-      allocated,
-      available: total - allocated,
-      utilizationPercent: total > 0 ? Math.round((allocated / total) * 100) : 0 }
-  })()
-
-  // Get manufacturer breakdown
-  const manufacturerBreakdown = (() => {
-    const mfgMap = new Map<string, number>()
-    gpuTypeInfo.forEach(info => {
-      const existing = mfgMap.get(info.manufacturer) || 0
-      mfgMap.set(info.manufacturer, existing + info.totalGPUs)
-    })
-    return Array.from(mfgMap.entries()).sort((a, b) => b[1] - a[1])
-  })()
-
-  // Get GPU specifications from nodes (memory, family, CUDA version)
-  const gpuSpecs = useMemo(() => {
-    const specs = {
-      totalMemoryGB: 0,
-      families: new Set<string>(),
-      cudaDriverVersions: new Set<string>(),
-      cudaRuntimeVersions: new Set<string>(),
-      migCapableCount: 0 }
-
-    gpuNodes.forEach(node => {
-      if (node.gpuMemoryMB) {
-        specs.totalMemoryGB += (node.gpuMemoryMB / 1024) * node.gpuCount
-      }
-      if (node.gpuFamily) {
-        specs.families.add(node.gpuFamily)
-      }
-      if (node.cudaDriverVersion) {
-        specs.cudaDriverVersions.add(node.cudaDriverVersion)
-      }
-      if (node.cudaRuntimeVersion) {
-        specs.cudaRuntimeVersions.add(node.cudaRuntimeVersion)
-      }
-      if (node.migCapable) {
-        specs.migCapableCount += node.gpuCount
-      }
-    })
-
-    return {
-      totalMemoryGB: Math.round(specs.totalMemoryGB),
-      families: Array.from(specs.families),
-      cudaDriverVersions: Array.from(specs.cudaDriverVersions),
-      cudaRuntimeVersions: Array.from(specs.cudaRuntimeVersions),
-      migCapableCount: specs.migCapableCount }
-  }, [gpuNodes])
+  const hasOperatorStatus = operatorStatus && operatorStatus.length > 0 && operatorStatus.some(s => s.gpuOperator || s.networkOperator)
+  const hasSpecs = gpuSpecs.totalMemoryGB > 0 || gpuSpecs.families.length > 0 || gpuSpecs.cudaDriverVersions.length > 0
 
   return (
     <BaseModal isOpen={isOpen} onClose={onClose} size="lg">
@@ -230,92 +95,24 @@ export function GPUDetailModal({ isOpen = true, gpuNodes, isLoading, error, onRe
           </div>
 
           {/* GPU Specifications */}
-          {(gpuSpecs.totalMemoryGB > 0 || gpuSpecs.families.length > 0 || gpuSpecs.cudaDriverVersions.length > 0) && (
+          {hasSpecs && (
             <div>
               <h4 className="text-sm font-medium text-muted-foreground mb-3 flex items-center gap-2">
                 <CircuitBoard className="w-4 h-4" />
                 GPU Specifications
               </h4>
-              <div className="grid grid-cols-2 md:grid-cols-4 gap-3">
-                {gpuSpecs.totalMemoryGB > 0 && (
-                  <div className="glass p-3 rounded-lg">
-                    <div className="flex items-center gap-2 text-muted-foreground text-xs mb-1">
-                      <HardDrive className="w-3 h-3" />
-                      {wrapAbbreviations('Total VRAM')}
-                    </div>
-                    <div className="text-lg font-bold text-foreground">{gpuSpecs.totalMemoryGB} GB</div>
-                  </div>
-                )}
-                {gpuSpecs.families.length > 0 && (
-                  <div className="glass p-3 rounded-lg">
-                    <div className="flex items-center gap-2 text-muted-foreground text-xs mb-1">
-                      <Cpu className="w-3 h-3" />
-                      Architecture
-                    </div>
-                    <div className="text-sm font-medium text-foreground capitalize">
-                      {(gpuSpecs.families || []).join(', ')}
-                    </div>
-                  </div>
-                )}
-                {gpuSpecs.cudaDriverVersions.length > 0 && (
-                  <div className="glass p-3 rounded-lg">
-                    <div className="flex items-center gap-2 text-muted-foreground text-xs mb-1">
-                      <Settings className="w-3 h-3" />
-                      {wrapAbbreviations('CUDA Driver')}
-                    </div>
-                    <div className="text-sm font-medium text-foreground">
-                      {(gpuSpecs.cudaDriverVersions || []).join(', ')}
-                    </div>
-                  </div>
-                )}
-                {gpuSpecs.migCapableCount > 0 && (
-                  <div className="glass p-3 rounded-lg">
-                    <div className="flex items-center gap-2 text-muted-foreground text-xs mb-1">
-                      <Layers className="w-3 h-3" />
-                      {wrapAbbreviations('MIG Capable')}
-                    </div>
-                    <div className="text-lg font-bold text-purple-400">{gpuSpecs.migCapableCount}</div>
-                  </div>
-                )}
-              </div>
+              <GPUSpecsPanel gpuSpecs={gpuSpecs} />
             </div>
           )}
 
           {/* NVIDIA Operator Status */}
-          {operatorStatus && operatorStatus.length > 0 && operatorStatus.some(s => s.gpuOperator || s.networkOperator) && (
+          {hasOperatorStatus && (
             <div>
               <h4 className="text-sm font-medium text-muted-foreground mb-3 flex items-center gap-2">
                 <Settings className="w-4 h-4" />
                 NVIDIA Operators
               </h4>
-              <div className="space-y-2">
-                {operatorStatus.map(status => (
-                  <div key={status.cluster} className="glass p-3 rounded-lg">
-                    <div className="text-sm font-medium text-foreground mb-2">{status.cluster}</div>
-                    <div className="flex flex-wrap gap-2">
-                      {status.gpuOperator && (
-                        <span className={`px-2 py-1 rounded text-xs ${
-                          status.gpuOperator.ready
-                            ? 'bg-green-500/20 text-green-400'
-                            : 'bg-yellow-500/20 text-yellow-400'
-                        }`}>
-                          GPU Operator: {status.gpuOperator.state || (status.gpuOperator.ready ? 'Ready' : 'Not Ready')}
-                          {status.gpuOperator.driverVersion && ` (${status.gpuOperator.driverVersion})`}
-                        </span>
-                      )}
-                      {status.networkOperator && (
-                        <span className={`px-2 py-1 rounded text-xs ${
-                          status.networkOperator.ready
-                            ? 'bg-green-500/20 text-green-400'
-                            : 'bg-yellow-500/20 text-yellow-400'
-                        }`}>
-                          Network Operator: {status.networkOperator.state || (status.networkOperator.ready ? 'Ready' : 'Not Ready')}
-                        </span>
-                      )}
-                    </div>
-                  </div>
-                ))}
-              </div>
+              <NvidiaOperatorStatusPanel operatorStatus={operatorStatus!} />
             </div>
           )}
 
@@ -351,35 +148,7 @@ export function GPUDetailModal({ isOpen = true, gpuNodes, isLoading, error, onRe
                 <Zap className="w-4 h-4" />
                 GPU Types
               </h4>
-              <div className="space-y-2">
-                {gpuTypeInfo.map(info => {
-                  const utilPercent = info.totalGPUs > 0 ? Math.round((info.allocatedGPUs / info.totalGPUs) * 100) : 0
-                  return (
-                    <div key={info.type} className="glass p-3 rounded-lg">
-                      <div className="flex items-center justify-between mb-2">
-                        <span className="font-medium text-foreground">{info.type}</span>
-                        <span className={`text-sm ${getUtilizationColor(utilPercent)}`}>
-                          {info.allocatedGPUs}/{info.totalGPUs} ({utilPercent}%)
-                        </span>
-                      </div>
-                      <div className="h-2 bg-secondary rounded-full overflow-hidden">
-                        <div
-                          className={`h-full transition-all ${
-                            utilPercent >= GPU_UTIL_HIGH ? 'bg-red-400' :
-                            utilPercent >= GPU_UTIL_WARN ? 'bg-yellow-400' :
-                            'bg-green-400'
-                          }`}
-                          style={{ width: `${utilPercent}%` }}
-                        />
-                      </div>
-                      <div className="flex items-center justify-between mt-2 text-xs text-muted-foreground">
-                        <span>{info.nodeCount} node{info.nodeCount !== 1 ? 's' : ''}</span>
-                        <span>{info.clusters.length} cluster{info.clusters.length !== 1 ? 's' : ''}: {info.clusters.join(', ')}</span>
-                      </div>
-                    </div>
-                  )
-                })}
-              </div>
+              <GPUTypeUtilizationList gpuTypeInfo={gpuTypeInfo} />
             </div>
           )}
 
@@ -390,35 +159,7 @@ export function GPUDetailModal({ isOpen = true, gpuNodes, isLoading, error, onRe
                 <Layers className="w-4 h-4" />
                 By Cluster
               </h4>
-              <div className="space-y-2">
-                {clusterInfo.map(info => {
-                  const utilPercent = info.totalGPUs > 0 ? Math.round((info.allocatedGPUs / info.totalGPUs) * 100) : 0
-                  return (
-                    <div key={info.cluster} className="glass p-3 rounded-lg">
-                      <div className="flex items-center justify-between mb-2">
-                        <span className="font-medium text-foreground">{info.cluster}</span>
-                        <span className={`text-sm ${getUtilizationColor(utilPercent)}`}>
-                          {info.allocatedGPUs}/{info.totalGPUs} allocated
-                        </span>
-                      </div>
-                      <div className="h-2 bg-secondary rounded-full overflow-hidden">
-                        <div
-                          className={`h-full transition-all ${
-                            utilPercent >= GPU_UTIL_HIGH ? 'bg-red-400' :
-                            utilPercent >= GPU_UTIL_WARN ? 'bg-yellow-400' :
-                            'bg-green-400'
-                          }`}
-                          style={{ width: `${utilPercent}%` }}
-                        />
-                      </div>
-                      <div className="flex items-center justify-between mt-2 text-xs text-muted-foreground">
-                        <span>{info.nodeCount} GPU node{info.nodeCount !== 1 ? 's' : ''}</span>
-                        <span>{(info.gpuTypes || []).join(', ')}</span>
-                      </div>
-                    </div>
-                  )
-                })}
-              </div>
+              <GPUClusterUtilizationList clusterInfo={clusterInfo} />
             </div>
           )}
 
@@ -429,57 +170,7 @@ export function GPUDetailModal({ isOpen = true, gpuNodes, isLoading, error, onRe
                 <Server className="w-4 h-4" />
                 GPU Nodes ({gpuNodes.length})
               </h4>
-              <div className="overflow-x-auto">
-                <table className="w-full text-sm">
-                  <thead>
-                    <tr className="border-b border-border">
-                      <th className="text-left py-2 px-3 text-muted-foreground font-medium">{t('common.node')}</th>
-                      <th className="text-left py-2 px-3 text-muted-foreground font-medium">{t('common.cluster')}</th>
-                      <th className="text-left py-2 px-3 text-muted-foreground font-medium">GPU Type</th>
-                      <th className="text-center py-2 px-3 text-muted-foreground font-medium">{t('common.memory')}</th>
-                      <th className="text-center py-2 px-3 text-muted-foreground font-medium">{t('common.used')}</th>
-                      <th className="text-center py-2 px-3 text-muted-foreground font-medium">{t('common.available')}</th>
-                      <th className="text-center py-2 px-3 text-muted-foreground font-medium">{t('common.total')}</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {gpuNodes.map(node => {
-                      const available = node.gpuCount - node.gpuAllocated
-                      const utilPercent = node.gpuCount > 0 ? Math.round((node.gpuAllocated / node.gpuCount) * 100) : 0
-                      const memoryGB = node.gpuMemoryMB ? Math.round(node.gpuMemoryMB / 1024) : null
-                      return (
-                        <tr key={`${node.cluster}-${node.name}`} className="border-b border-border/50 hover:bg-secondary/30">
-                          <td className="py-2 px-3 font-mono text-xs text-foreground">
-                            <div className="flex items-center gap-1">
-                              {node.name}
-                              {node.migCapable && (
-                                <StatusBadge color="purple" size="xs">MIG</StatusBadge>
-                              )}
-                            </div>
-                          </td>
-                          <td className="py-2 px-3 text-muted-foreground">{node.cluster}</td>
-                          <td className="py-2 px-3 text-muted-foreground">
-                            <div>
-                              {node.gpuType}
-                              {node.gpuFamily && (
-                                <span className="text-xs text-muted-foreground/70 ml-1 capitalize">({node.gpuFamily})</span>
-                              )}
-                            </div>
-                          </td>
-                          <td className="py-2 px-3 text-center text-muted-foreground">
-                            {memoryGB ? `${memoryGB}GB` : '-'}
-                          </td>
-                          <td className={`py-2 px-3 text-center ${getUtilizationColor(utilPercent)}`}>
-                            {node.gpuAllocated}
-                          </td>
-                          <td className="py-2 px-3 text-center text-green-400">{available}</td>
-                          <td className="py-2 px-3 text-center text-foreground">{node.gpuCount}</td>
-                        </tr>
-                      )
-                    })}
-                  </tbody>
-                </table>
-              </div>
+              <GPUNodesTable gpuNodes={gpuNodes} />
             </div>
           )}
